// src/app/generated/models/financial-report.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FinancialPartialData } from '../models/financial-partial-data';
import { FinancialReportPeriod } from '../models/financial-report-period';
import { PeriodType } from '../models/period-type';
export interface FinancialReport {
  endDate?: string;
  periodType?: PeriodType;
  periods?: Array<FinancialReportPeriod>;
  startDate?: string;
  totalBalance?: FinancialPartialData;
  totalExpense?: FinancialPartialData;
  totalIncome?: FinancialPartialData;
}