// src/app/generated/models/financial-report-period.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FinancialPartialData } from '../models/financial-partial-data';
import { FinancialReportCategoryGroup } from '../models/financial-report-category-group';
export interface FinancialReportPeriod {
  balance?: FinancialPartialData;
  endDate?: string;
  expense?: FinancialReportCategoryGroup;
  income?: FinancialReportCategoryGroup;
  startDate?: string;
}